backup_before_refactor/requirements.txt
# Core dependencies - 数据管道版本
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0

# Async support - 异步任务调度
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Data processing
ccxt>=4.0.0
web3>=6.0.0

# Workflow scheduling - 可选，用于生产环境
# apache-airflow>=2.5.0

# Database migration - 数据库迁移
alembic>=1.8.0

# Utilities
python-dotenv>=0.19.0
loguru>=0.6.0
pydantic>=1.9.0

# Development - 保留基础工具
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
faker>=15.0.0
freezegun>=1.2.0
